+++ src/inspect_ai/scorer/_metrics/__init__.py
@@ -1,10 +1,11 @@
 from .accuracy import accuracy
-from .mean import mean
+from .mean import grouped_mean, mean
 from .std import bootstrap_stderr, std, stderr, var
 
 __all__ = [
     "accuracy",
     "mean",
+    "grouped_mean",
     "bootstrap_stderr",
     "std",
     "stderr",