+++ tests/scorer/test_metric.py
@@ -1,5 +1,7 @@
 from typing import Any, Callable, cast
 
+import pytest
+
 from inspect_ai import Task, eval, score
 from inspect_ai._util.constants import PKG_NAME
 from inspect_ai._util.registry import registry_info
@@ -10,6 +12,7 @@ from inspect_ai.scorer import (
     Scorer,
     Value,
     accuracy,
+    grouped_mean,
     includes,
     match,
     mean,
@@ -382,3 +385,48 @@ def test_clustered_stderr():
         ]
     )
     assert round(se, 3) == 0.645
+
+
+def test_grouped_mean_single():
+    metric = grouped_mean(group_key="group")
+    result = metric(
+        [
+            SampleScore(score=Score(value=1), sample_metadata={"group": "A"}),
+            SampleScore(score=Score(value=1), sample_metadata={"group": "A"}),
+            SampleScore(score=Score(value=4), sample_metadata={"group": "A"}),
+        ]
+    )
+    assert result["A"] == 2.0
+    assert result["all"] == 2.0
+
+
+def test_grouped_mean_multiple():
+    metric = grouped_mean(group_key="group")
+    result = metric(
+        [
+            SampleScore(score=Score(value=1), sample_metadata={"group": "A"}),
+            SampleScore(score=Score(value=1), sample_metadata={"group": "A"}),
+            SampleScore(score=Score(value=4), sample_metadata={"group": "A"}),
+            SampleScore(score=Score(value=2), sample_metadata={"group": "B"}),
+            SampleScore(score=Score(value=6), sample_metadata={"group": "B"}),
+            SampleScore(score=Score(value=10), sample_metadata={"group": "B"}),
+        ]
+    )
+    assert result["A"] == 2.0
+    assert result["B"] == 6.0
+    assert result["all"] == 4.0
+
+
+def test_grouped_mean_error():
+    metric = grouped_mean(group_key="group")
+    with pytest.raises(ValueError):
+        metric(
+            [
+                SampleScore(score=Score(value=1), sample_metadata={"group": "A"}),
+                SampleScore(score=Score(value=1)),
+                SampleScore(score=Score(value=4), sample_metadata={"group": "A"}),
+                SampleScore(score=Score(value=2), sample_metadata={"group": "B"}),
+                SampleScore(score=Score(value=6), sample_metadata={"group": "B"}),
+                SampleScore(score=Score(value=10), sample_metadata={"group": "B"}),
+            ]
+        )